+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## x.x.x
 * Improve loading time for native ads. (https://github.com/AppLovin/AppLovin-MAX-React-Native/issues/273)
+* Depend on Android SDK 12.1.0 and iOS SDK 12.1.0.
 ## 6.1.0
 * Fix inconsistent naming for enum fields.
 * Add support for Terms and Privacy Policy Flow. For more info, check out our [docs](https://dash.applovin.com/documentation/mediation/react-native/getting-started/terms-and-privacy-policy-flow).