+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Versions
 
 ## x.x.x
+* Improve loading time for native ads. (https://github.com/AppLovin/AppLovin-MAX-React-Native/issues/273)
 * Depend on Android SDK 12.1.0 and iOS SDK 12.1.0.
 ## 6.1.0
 * Fix inconsistent naming for enum fields.